requirements.txt
Django==5.2.1
djangorestframework==3.16.0
mysqlclient==2.2.7
Pillow==10.2.0
coreapi==2.3.3
django-filter==23.5
django-cors-headers==4.3.1 